Addition of new format needed for ShakeCast

+++ tests/autogf_test.py
@@ -33,39 +33,44 @@ def test_autogf(tmpdir):
     """ % (datadir, upone, upone, datadir)
     # Make a copy of current defaults
     default_file = os.path.join(os.path.expanduser("~"), ".gfail_defaults")
-    shutil.copy(default_file, default_file+'_bak')
+    if os.path.exists(default_file):
+        shutil.copy(default_file, default_file+'_bak')
 
-    #p = os.path.join(str(tmpdir.name), "sub")
-    p = os.path.join(str(tmpdir), "sub")
-    if not os.path.exists(p):
-        os.makedirs(p)
+    try:
+        #p = os.path.join(str(tmpdir.name), "sub")
+        p = os.path.join(str(tmpdir), "sub")
+        if not os.path.exists(p):
+            os.makedirs(p)
 
-    # Modify gfail defaults
-    pathcmd = pathcmd.replace('[TMPOUT]', p)
-    rc, so, se = get_command_output(pathcmd)
+        # Modify gfail defaults
+        pathcmd = pathcmd.replace('[TMPOUT]', p)
+        rc, so, se = get_command_output(pathcmd)
 
-    # Set command for autogf
-    agfcmd = """
-        autogf -c %s \
-        -d test.db \
-        -a %s \
-        -t -w
-    """ % (gconfig, config)
-    
-    rc, so, se = get_command_output(agfcmd)
-    temp = so.decode().split('\n')[-2]
-    
+        # Set command for autogf
+        agfcmd = """
+            autogf -c %s \
+            -d test.db \
+            -a %s \
+            -t -w
+        """ % (gconfig, config)
+
+        rc, so, se = get_command_output(agfcmd)
+        temp = so.decode().split('\n')[-2]
+    except Exception as e:  # To make sure defaults are replaced
+        print(e)
+
+    # Change back gfail defaults
+    if os.path.exists(default_file+'_bak'):
+        shutil.copy(default_file+'_bak', default_file)
+    # Remove backup and tempfile
+    os.remove(default_file+'_bak')
+    shutil.rmtree(p)
+
+    # Then run tests
     np.testing.assert_equal(True, rc, 'autogf run failed')
     np.testing.assert_string_equal(temp, 'Test successful, cleaning up files')
     np.testing.assert_equal(True, os.path.exists('test.db'))
     os.remove('test.db')
-    
-    # Change back gfail defaults
-    shutil.copy(default_file+'_bak', default_file)
-
-    # Remove backup and tempfile
-    os.remove(default_file+'_bak')
-    shutil.rmtree(p)
 
 
 if __name__ == "__main__":
